+++ .vscode/launch.json
@@ -183,6 +183,7 @@
         "morpheus-swap",
         "morphex",
         "morpho-aave",
+        "morpho-aavev3",
         "morpho-compound",
         "mstable",
         "multichain",
