+++ src/adapters/index.ts
@@ -114,6 +114,7 @@ import metronome from '@adapters/metronome'
 import morpheusSwap from '@adapters/morpheus-swap'
 import morphex from '@adapters/morphex'
 import morphoAave from '@adapters/morpho-aave'
+import morphoAavev3 from '@adapters/morpho-aavev3'
 import morphoCompound from '@adapters/morpho-compound'
 import mstable from '@adapters/mstable'
 import multichain from '@adapters/multichain'
@@ -336,6 +337,7 @@ export const adapters: Adapter[] = [
   morpheusSwap,
   morphex,
   morphoAave,
+  morphoAavev3,
   morphoCompound,
   mstable,
   multichain,
