+++ src/adapters/morpho-aavev3/ethereum/index.ts
@@ -1,16 +1,15 @@
-import { getMarketsContracts } from '@adapters/morpho-aave/ethereum/contract'
+import { getLendBorrowBalancesAaveV3, getUserHealthFactorV3 } from '@adapters/morpho-aavev3/ethereum/balance'
+import { getMarketsContractsMorphoAaveV3 } from '@adapters/morpho-aavev3/ethereum/contract'
 import type { BaseContext, Contract, GetBalancesHandler } from '@lib/adapter'
 import { resolveBalances } from '@lib/balance'
 
-import { getLendBorrowBalances, getUserHealthFactor } from './balances'
-
-const lens: Contract = {
+const morphoAave3Proxy: Contract = {
   chain: 'ethereum',
-  address: '0x507fa343d0a90786d86c7cd885f5c49263a91ff4',
+  address: '0x33333aea097c193e66081e930c33020272b33333',
 }
 
 export const getContracts = async (ctx: BaseContext) => {
-  const markets = await getMarketsContracts(ctx, lens)
+  const markets = await getMarketsContractsMorphoAaveV3(ctx, morphoAave3Proxy)
 
   return {
     contracts: { markets },
@@ -20,9 +19,9 @@ export const getContracts = async (ctx: BaseContext) => {
 export const getBalances: GetBalancesHandler<typeof getContracts> = async (ctx, contracts) => {
   const [balances, healthFactor] = await Promise.all([
     resolveBalances<typeof getContracts>(ctx, contracts, {
-      markets: (...args) => getLendBorrowBalances(...args, lens),
+      markets: (...args) => getLendBorrowBalancesAaveV3(...args, morphoAave3Proxy),
     }),
-    getUserHealthFactor(ctx, lens),
+    getUserHealthFactorV3(ctx, morphoAave3Proxy),
   ])
 
   return {
